Implement authentication with AuthController; add user registration, login, and logout methods; include middleware for pedido ownership

+++ routes/api.php
@@ -1,28 +1,37 @@
 <?php
 
+use App\Http\Controllers\AuthController;
+use App\Http\Controllers\PedidosController;
+use App\Http\Controllers\ProdutosController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
-use App\Http\Controllers\ProdutosController;
-use App\Http\Controllers\PedidosController;
+
 
 Route::get('/user', function (Request $request) {
     return $request->user();
 })->middleware('auth:sanctum');
 
+Route::post('/register', [AuthController::class, 'register'])->name('register');
+Route::post('/login', [AuthController::class, 'login'])->name('login');
+Route::post('/logout', [AuthController::class, 'logout'])->middleware('auth:sanctum')->name('logout');
 
-Route::prefix('produtos')->group(function () {
-    Route::get('/', [ProdutosController::class, 'index']);
-    Route::post('/', [ProdutosController::class, 'store']);
-    Route::get('/{id}', [ProdutosController::class, 'show']);
-    Route::put('/{id}', [ProdutosController::class, 'update']);
-    Route::delete('/{id}', [ProdutosController::class, 'destroy']);
+
+Route::middleware('auth:sanctum')->group(function () {  
+
+    Route::prefix('produtos')->group(function () {
+        Route::get('/', [ProdutosController::class, 'index']);
+        Route::post('/', [ProdutosController::class, 'store']);
+        Route::get('/{id}', [ProdutosController::class, 'show']);
+        Route::put('/{id}', [ProdutosController::class, 'update']);
+        Route::delete('/{id}', [ProdutosController::class, 'destroy']);
+    });
+
+    Route::prefix('pedidos')->group(function () {
+        Route::get('/', [PedidosController::class, 'index']);
+        Route::post('/', [PedidosController::class, 'store']);
+        Route::get('/{id}', [PedidosController::class, 'show']);
+        Route::put('/{id}', [PedidosController::class, 'update'])->middleware('pedido.owner');
+        Route::delete('/{id}', [PedidosController::class, 'destroy'])->middleware('pedido.owner');
+    });
+
 });
-
-Route::prefix('pedidos')->group(function () {
-    Route::get('/', [PedidosController::class, 'index']);
-    Route::post('/', [PedidosController::class, 'store']);
-    Route::get('/{id}', [PedidosController::class, 'show']);
-    Route::put('/{id}', [PedidosController::class, 'update']);
-    Route::delete('/{id}', [PedidosController::class, 'destroy']);
-});
-
